Seperated sleep part of process, to allow multithreaded applications
To do this, a lock should be held while processTasklets and
processDrivers are called. The lock can be released when sleep is
called. The lock should also be used to protect calls to the
openthread library outside of the openthread context. (Callbacks
from openthread are in the openthread context).

Example in progress...