Improve program structure flexibility and fix issues (#12)
* Prevent double-initialisation of uart

* typo

* Updated ot

* Remove now-redundant code

* updated ot

* Seperated sleep part of process, to allow multithreaded applications

To do this, a lock should be held while processTasklets and
processDrivers are called. The lock can be released when sleep is
called. The lock should also be used to protect calls to the
openthread library outside of the openthread context. (Callbacks
from openthread are in the openthread context).

Example in progress...

* Slight restructure and add multithreaded example

* Reset makefile to just make main, TODO: fix later

* Improve flexibility of the platform api to allow different program flows

* fixed pointer error

* Quick fix to speed up flash emulation at startup -> replace with non-emulation settings layer

* Maintain a consistent IeeeEui64 during a single run