Use the NODE_ID in file names to enable multiple ot processes (#45)
This enables instancing of openthread nodes. Currently, the global NODE_ID variable must be set to a unique value for each process, which affects which config files the process uses. The example program does this by passing in a command line argument. In future this should be made more dynamic.